pyproject.toml
[project]
name = "purple-rooms-solver"
version = "0.1.0"
description = "Baseline purple agent for Rooms benchmark"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "pydantic>=2.12.5",
    "uvicorn>=0.38.0",
]

# --- ADD THIS SECTION ---
[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# --- ADD THIS SECTION ---
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
asyncio_default_fixture_loop_scope = "function"